// src/api/models/MaterialDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PK.PHCalc.OutsourcingDummy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MaterialDTO
 */
export interface MaterialDTO {
    /**
     * 
     * @type {string}
     * @memberof MaterialDTO
     */
    prowellName: string;
    /**
     * 
     * @type {number}
     * @memberof MaterialDTO
     */
    minimalTotalAreaInSquareMeters: number;
    /**
     * 
     * @type {number}
     * @memberof MaterialDTO
     */
    price: number;
}

export function MaterialDTOFromJSON(json: any): MaterialDTO {
    return MaterialDTOFromJSONTyped(json, false);
}

export function MaterialDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): MaterialDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'prowellName': json['ProwellName'],
        'minimalTotalAreaInSquareMeters': json['MinimalTotalAreaInSquareMeters'],
        'price': json['Price'],
    };
}

export function MaterialDTOToJSON(value?: MaterialDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'ProwellName': value.prowellName,
        'MinimalTotalAreaInSquareMeters': value.minimalTotalAreaInSquareMeters,
        'Price': value.price,
    };
}